# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /worker/jobs/{subId}:
    get:
      operationId: getWorkerJobs
      tags: [worker]
      summary: Get all jobs assigned to a specific subcontractor
      parameters:
        - in: path
          name: subId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assigned jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /worker/assignments/{id}/status:
    patch:
      operationId: workerUpdateAssignmentStatus
      tags: [worker]
      summary: Worker updates their assignment status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "404":
          description: Not found
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        serviceType:
          type: string
        planName:
          type: ["string", "null"]
        customerPrice:
          type: ["number", "null"]
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        assignment:
          $ref: "#/components/schemas/Assignment"
          nullable: true
      required:
        - id
        - customerName
        - phone
        - address
        - serviceType
        - status
        - createdAt
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        subcontractorId:
          type: integer
        subName:
          type: ["string", "null"]
        subPay:
          type: number
        ownerProfit:
          type: number
        status:
          type: string
        assignedAt:
          type: string
        completedAt:
          type: ["string", "null"]
      required:
        - id
        - jobId
        - subcontractorId
        - subPay
        - ownerProfit
        - status
        - assignedAt
...
    UpdateAssignmentBody:
      type: object
      properties:
        status:
          type: string
        completedAt:
          type: ["string", "null"]
...